backend/requirements.txt
Django==5.2.5
django-ninja==1.4.3
django-cors-headers==4.7.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
uvicorn[standard]==0.35.0
django-modeltranslation==0.19.16
whitenoise>=6.5,<7
django-elasticsearch-dsl==8.2.0
